Settle on "Troubleshoot Helper" as the add-on's name.  Update IDs, etc.
A couple of changes to whitelistedDOMAPI.  defineProperty should define the property on matching windows that are already open, not only on windows opened in the future.  Because of that, it needs to iterate over all content windows, so also make it accept an array of origins rather than a single origin so that it doesn't need to iterate over content windows multiple times.

Add a copy of Troubleshoot.jsm to the add-on so that people running Firefoxes older than 18 can use it.  Had to modify it to remove preprocessor instructions.

Set the minumum supported Firefox version to 10.  No idea if it actually works.